move data to MediaPlayer class
